These tuples.gz files contain EXACTLY the same content as prior.
Now, they are generated w/out embedded information about the current
date.  This means that if you redo the configuration for the testsuite,
the files will ONLY change if an actual change occurs to the testsuite,
not just a re-run of the existing configuration.

There are ZERO, repeat ZERO changes to the content of the files,
only to the gzip envelope.